+++ internal/dto/common/request.go
@@ -1,10 +1,8 @@
 package common
 
-import "book-manager/internal/models"
-
 type Request[T any] struct {
-	RequestId   string           `json:"request_id"`
-	RequestTime string           `json:"request_time"`
-	Data        T                `json:"data"`
-	Paginate    *models.Paginate `json:"paginate"`
+	RequestId   string    `json:"request_id"`
+	RequestTime string    `json:"request_time"`
+	Data        T         `json:"data"`
+	Paginate    *Paginate `json:"paginate"`
 }