+++ internal/services/impl/author_service_impl.go
@@ -4,7 +4,6 @@ import (
 	"book-manager/internal/dto/author"
 	"book-manager/internal/dto/common"
 	"book-manager/internal/mapper"
-	"book-manager/internal/models"
 	"book-manager/internal/repositories"
 	"book-manager/internal/services"
 	"book-manager/internal/utils"
@@ -17,7 +16,8 @@ type AuthorService struct {
 }
 
 func (a AuthorService) Delete(req *common.Request[author.DeleteAuthor]) common.Response[any] {
-	err := a.Repo.Delete(req, "Anonymous")
+	ok, err := a.Repo.Delete(req, "Anonymous")
+
 	if err != nil {
 		var appErr *error_codes.AppError
 		if errors.As(err, &appErr) {
@@ -28,11 +28,16 @@ func (a AuthorService) Delete(req *common.Request[author.DeleteAuthor]) common.R
 		}
 		return utils.BuildResponse[any](nil, error_codes.BadRequest, req.RequestId)
 	}
+
+	if !ok {
+		return utils.BuildResponse[any](nil, error_codes.BadRequest, req.RequestId)
+	}
+
 	return utils.BuildResponse[any](nil, error_codes.Success, req.RequestId)
 }
 
 func (a AuthorService) Update(req *common.Request[author.UpdateAuthor]) common.Response[any] {
-	err := a.Repo.Update(req, "Anonymous")
+	ok, err := a.Repo.Update(req, "Anonymous")
 
 	if err != nil {
 		var appErr *error_codes.AppError
@@ -44,45 +49,70 @@ func (a AuthorService) Update(req *common.Request[author.UpdateAuthor]) common.R
 		}
 		return utils.BuildResponse[any](nil, error_codes.BadRequest, req.RequestId)
 	}
+
+	if !ok {
+		return utils.BuildResponse[any](nil, error_codes.BadRequest, req.RequestId)
+	}
+
 	return utils.BuildResponse[any](nil, error_codes.Success, req.RequestId)
 }
 
-func (a AuthorService) GetAll(req *common.Request[any]) common.Response[[]models.Author] {
+func (a AuthorService) GetAll(req *common.Request[any]) common.Response[[]author.AuthorResponse] {
 	data, err := a.Repo.GetAll(req)
+
 	if err != nil {
 		var appErr *error_codes.AppError
 		if errors.As(err, &appErr) {
-			return utils.BuildResponse[[]models.Author]([]models.Author{}, error_codes.ErrorCode{
+			return utils.BuildResponse[[]author.AuthorResponse](nil, error_codes.ErrorCode{
 				Code: appErr.Code,
 				Msg:  appErr.Message,
 			}, req.RequestId)
 		}
-		return utils.BuildResponse[[]models.Author]([]models.Author{}, error_codes.BadRequest, req.RequestId)
+		return utils.BuildResponse[[]author.AuthorResponse](nil, error_codes.BadRequest, req.RequestId)
 	}
-	return utils.BuildResponse[[]models.Author](data, error_codes.Success, req.RequestId)
+
+	// Map entities to response DTOs
+	var result []author.AuthorResponse
+	for _, item := range data {
+		result = append(result, author.AuthorResponse{
+			Id:   item.Id,
+			Name: item.Name,
+		})
+	}
+
+	return utils.BuildResponse[[]author.AuthorResponse](result, error_codes.Success, req.RequestId)
 }
 
-func (a AuthorService) GetOne(req *common.Request[author.GetOneAuthor]) common.Response[models.Author] {
+func (a AuthorService) GetOne(req *common.Request[author.GetOneAuthor]) common.Response[author.AuthorResponse] {
+
 	data, err := a.Repo.GetOne(req)
 
 	if err != nil {
 		var appErr *error_codes.AppError
 		if errors.As(err, &appErr) {
-			return utils.BuildResponse[models.Author](models.Author{}, error_codes.ErrorCode{
+			return utils.BuildResponse[author.AuthorResponse](author.AuthorResponse{}, error_codes.ErrorCode{
 				Code: appErr.Code,
 				Msg:  appErr.Message,
 			}, req.RequestId)
 		}
-		return utils.BuildResponse[models.Author](models.Author{}, error_codes.BadRequest, req.RequestId)
+		return utils.BuildResponse[author.AuthorResponse](author.AuthorResponse{}, error_codes.BadRequest, req.RequestId)
 	}
-	return utils.BuildResponse[models.Author](data, error_codes.Success, req.RequestId)
+
+	// Map entity to response DTO
+	result := author.AuthorResponse{
+		Id:   data.Id,
+		Name: data.Name,
+	}
+
+	return utils.BuildResponse[author.AuthorResponse](result, error_codes.Success, req.RequestId)
 }
 
 func (a AuthorService) Create(req *common.Request[author.AddAuthor]) common.Response[any] {
 	authorEntity := mapper.AuthorMapper(req.Data, "Anonymous")
+
 	request := mapper.RequestMapper(req, authorEntity)
 
-	err := a.Repo.Create(request)
+	ok, err := a.Repo.Create(request)
 
 	if err != nil {
 		var appErr *error_codes.AppError
@@ -92,11 +122,14 @@ func (a AuthorService) Create(req *common.Request[author.AddAuthor]) common.Resp
 				Msg:  appErr.Message,
 			}, req.RequestId)
 		}
-
 		return utils.BuildResponse[any](nil, error_codes.BadRequest, req.RequestId)
 	}
-	return utils.BuildResponse[any](nil, error_codes.Success, req.RequestId)
 
+	if !ok {
+		return utils.BuildResponse[any](nil, error_codes.BadRequest, req.RequestId)
+	}
+
+	return utils.BuildResponse[any](nil, error_codes.Success, req.RequestId)
 }
 
 func NewAuthorService(repo repositories.IAuthorRepository) services.IAuthorService {